// server/src/domain/enums/SocialPlatform.ts
export enum SocialPlatform {
  FACEBOOK = 'FACEBOOK',
  TWITTER = 'TWITTER', // X olarak da bilinir
  LINKEDIN = 'LINKEDIN',
  INSTAGRAM = 'INSTAGRAM',
  YOUTUBE = 'YOUTUBE',
  TIKTOK = 'TIKTOK',
  // Eklenen yeni sosyal medya platformları
  PINTEREST = 'PINTEREST',
  SNAPCHAT = 'SNAPCHAT',
  REDDIT = 'REDDIT',
  DISCORD = 'DISCORD',
  TELEGRAM = 'TELEGRAM',
  WHATSAPP = 'WHATSAPP',
  UNKNOWN = 'UNKNOWN'
}